// desktop/src/types/mutations/node-attribute-set.ts
export type NodeAttributeSetMutationInput = {
  type: 'node_attribute_set';
  userId: string;
  nodeId: string;
  attribute: string;
  value: any;
};

export type NodeAttributeSetMutationOutput = {
  success: boolean;
};

declare module '@/types/mutations' {
  interface MutationMap {
    node_attribute_set: {
      input: NodeAttributeSetMutationInput;
      output: NodeAttributeSetMutationOutput;
    };
  }
}